requirements.txt
ccxt
schedule
pandas
python-binance
secure-smtplib
emails